feat(docs): move Header into root layout
