+++ app/docs/src/app/page.tsx
@@ -2,19 +2,10 @@ import { center } from "@styled-system/patterns"
 import Link from "next/link"
 
 import { Button } from "@/components/button"
-import { Header } from "@/components/header"
 
 export default function MainPage() {
-  const navItems = [
-    { label: "Home", href: "/" },
-    { label: "Components", href: "/components" },
-    { label: "Tokens", href: "/tokens" },
-    { label: "Guidelines", href: "/guidelines" },
-  ]
-
   return (
     <>
-      <Header logo={<span>Design System</span>} navItems={navItems} />
       <main
         className={center({
           flexGrow: "1",
